+++ src/djlint/settings.py
@@ -988,12 +988,12 @@ class Config:
             (?:
                 (
                     (?:
-                        (?:\w|-|\.|\:|@|/(?!>)) # a name character
+                        (?:\w|-|\.|\:|@|\*|/(?!>)) # a name character
                        | (?>{{{{[\s\S]*?}}}})
-                         (?=(?:\w|-|\.|\:|@|/(?!>))|[ ]*=) # a leading template variable
+                         (?=(?:\w|-|\.|\:|@|\*|/(?!>))|[ ]*=) # a leading template variable
                     )
                     (?:
-                        (?:\w|-|\.|\:|@|/(?!>)) # more name characters
+                        (?:\w|-|\.|\:|@|\*|/(?!>)) # more name characters
                        | (?>{{{{[\s\S]*?}}}}|{{%[\s\S]*?%}}) # or an embedded template tag
                     )*
                     | required | checked