+++ src/djlint/formatter/condense.py
@@ -273,7 +273,7 @@ def condense_html(html: str, config: Config, source: str | None = None) -> str:
         except StopIteration:
             was_authored_multiline = False
 
-        if was_authored_multiline:
+        if was_authored_multiline or inside_html_attribute(html, match):
             return match.group()
 
         return condense_line(config, html, match)